Helm/02-Centralize_System/requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
observer==0.1.0
PyYAML==6.0.2
requests==2.32.4
urllib3==2.5.0
watchdog==6.0.0